backend/pyproject.toml
[project]
name = "notes-backend"
version = "0.1.0"
description = "Django 5 + DRF backend for notes-challenge"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=2.3",
    "drf-spectacular>=0.27",
    "django-cors-headers>=4.3",
    "psycopg[binary]>=3.1",
    "gunicorn>=21.0",
    "python-dotenv>=1.0",
    "dj-database-url>=2.0",
    "Pillow>=10.0",
    "openai>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "core", "api"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "ruff>=0.4",
    "black>=24.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["config", "core", "api", "tests"]
exclude = ["migrations"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
)/
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.test_settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
